properties table

+++ branches/myawareness/adb/src/DatabaseConnection.cpp
@@ -1,114 +1,20 @@
-#include <cstdio>
 #include <Configuration.h>
 #include <Exception.h>
-#include <SelectionParameters.h>
 #include <DbUtil.h>
+#include <cmd/CreateDatabaseCommand.h>
 #include <cmd/ReversibleDatabaseCommand.h>
-#include <cmd/SelectTransactionsCommand.h>
+#include <cmd/SelectPreferences.h>
+#include <cmd/PurgeDatabaseCommand.h>
 #include <DatabaseConnection.h>
 
 using namespace std;
 
 namespace adb {
 
-    bool DatabaseConnection::isAccountInUse(sqlite3* database, int accountId)
+    DatabaseConnection::DatabaseConnection(const char* location) :
+        database_(0)
     {
-        SelectionParameters parameters;
-        parameters.setAccountId(accountId);
-        parameters.setCheckUsage(true);
-        vector<int> selection;
-        SelectTransactionsCommand(database, &selection, &parameters).execute();
-        return selection.size() > 0;
-
-    }
-
-    bool DatabaseConnection::isItemInUse(sqlite3* database, int itemId)
-    {
-        SelectionParameters parameters;
-        parameters.setItemId(itemId);
-        parameters.setCheckUsage(true);
-        vector<int> selection;
-        SelectTransactionsCommand(database, &selection, &parameters).execute();
-        return selection.size() > 0;
-    }
-
-    DatabaseConnection* DatabaseConnection::instance_ = 0;
-
-    bool DatabaseConnection::isOpened()
-    {
-        return 0 != instance_;
-    }
-
-    DatabaseConnection* DatabaseConnection::instance()
-    {
-        if (!isOpened()) {
-            openDatabase(Configuration::instance()->getLastDatabasePath().c_str());
-        }
-        return instance_;
-    }
-
-    void DatabaseConnection::openDatabase(const char* databasePath)
-    {
-        DatabaseConnection* tmpInstance = 0;
-        try {
-            tmpInstance = new DatabaseConnection(databasePath);
-            if (isOpened()) {
-                if (instance_->getDatabaseLocation() == databasePath) {
-                    return;
-                }
-                closeDatabase();
-            }
-            instance_ = tmpInstance;
-            Configuration::instance()->setLastDatabasePath(databasePath);
-        } catch (const exception& ex) {
-            delete tmpInstance;
-            RETHROW(ex);
-        }
-    }
-
-    void DatabaseConnection::closeDatabase()
-    {
-        if (!isOpened()) {
-            THROW(Exception::NO_DATABASE_MESSAGE);
-        }
-        delete instance_;
-        instance_ = 0;
-    }
-
-    void DatabaseConnection::deleteDatabase()
-    {
-        if (!isOpened()) {
-            THROW(Exception::NO_DATABASE_MESSAGE);
-        }
-        string databasePath = instance_->getDatabaseLocation();
-        closeDatabase();
-        if (0 != ::remove(databasePath.c_str())) {
-            THROW("error deleting database");
-        }
-    }
-
-    void DatabaseConnection::exportDatabase(ostream& out)
-    {
-        if (!isOpened()) {
-            THROW(Exception::NO_DATABASE_MESSAGE);
-        }
-        instance_->dumpSql(out);
-    }
-
-    void DatabaseConnection::importDatabase(istream& in)
-    {
-        if (!isOpened()) {
-            THROW(Exception::NO_DATABASE_MESSAGE);
-        }
-        string databasePath = instance_->getDatabaseLocation();
-        deleteDatabase();
-        openDatabase(databasePath.c_str());
-        instance_->loadSql(in);
-    }
-
-    DatabaseConnection::DatabaseConnection(const char* file) :
-        databaseLocation_(file), database_(0)
-    {
+        DbUtil::charPtrToString(databaseLocation_, location);
         DbUtil::trimSpaces(databaseLocation_);
         if (databaseLocation_.size() <= 0) {
             THROW("invalid file name");
@@ -118,41 +24,61 @@ namespace adb {
 
     DatabaseConnection::~DatabaseConnection()
     {
-        closeConnection();
+        closeConnection(true);
     }
 
     void DatabaseConnection::openConnection()
     {
         if (::sqlite3_open_v2(databaseLocation_.c_str(), &database_, SQLITE_OPEN_READWRITE | SQLITE_OPEN_CREATE, NULL)) {
-            closeConnection();
+            closeConnection(false);
             THROW("can't read database");
         }
 
-        switch (checkConnection()) {
-            case OK:
-                break;
+        int tableCount = getTableCount();
 
-            case EMPTY_DATABASE:
-                if (OK != createNewDatabase()) {
-                    THROW("error creating database");
-                }
-                break;
+        if (tableCount == 0) {
+            createNewDatabase();
+        } else if (tableCount > 0) {
+            SelectPreferences prefs(database_);
+            prefs.execute();
 
-            default:
-                closeConnection();
-                THROW("error reading database");
+            if (prefs[Configuration::PREF_PROJECT_MARKER] != Configuration::PROJECT_MARKER) {
+                closeConnection(false);
+                THROW("the database is not supported by this version of the application");
+            }
+
+            // TBD: check database version
+
+            readPreferences(prefs);
+        } else {
+            closeConnection(false);
+            THROW("error reading database");
         }
     }
 
-    int DatabaseConnection::checkConnection()
+    void DatabaseConnection::createNewDatabase()
     {
-        const char* zSql = "SELECT * FROM sqlite_master WHERE type='table';";
-        sqlite3_stmt* stmt = 0;
+        CreateDatabaseCommand(database_).execute();
+    }
 
+    void DatabaseConnection::purgeDatabase()
+    {
+        PurgeDatabaseCommand(database_).execute(); // TBD-: - optional via preferences
+        // TBD-: remove unused descriptions - optional via preferences
+        // TBD-: compact identical transactions in each day - optional via preferences
+        // TBD-: kill undo buffer
+    }
+
+    int DatabaseConnection::getTableCount()
+    {
+        // TBD-: use a command
+
+        sqlite3_stmt* stmt = 0;
+        const char* zSql = "SELECT * FROM sqlite_master WHERE type='table';";
         int rc = ::sqlite3_prepare_v2(database_, zSql, 1000, &stmt, NULL);
         if (SQLITE_OK != rc) {
             ::sqlite3_finalize(stmt);
-            return EMPTY_DATABASE;
+            return 0;
         }
 
         int tableCount = 0;
@@ -164,23 +90,22 @@ namespace adb {
                 break;
             } else {
                 ::sqlite3_finalize(stmt);
-                return STATEMENT_ERROR;
+                return -1;
             }
         }
-
-        if (0 == tableCount) {
-            return EMPTY_DATABASE;
-        }
-
-        return OK;
+        return tableCount;
     }
 
-    void DatabaseConnection::closeConnection()
+    void DatabaseConnection::closeConnection(bool purge)
     {
-        // TBD-: compact identical transactions in each day
-        purgeDatabase();
-        ::sqlite3_close(database_);
-        database_ = 0;
+        if (database_ != 0) {
+            if (purge) {
+                purgeDatabase();
+            }
+            writePreferences(database_);
+            ::sqlite3_close(database_);
+            database_ = 0;
+        }
     }
 
     void DatabaseConnection::undo()
